javascript highcharts and highstock updated to new versions
