+++ src/main/java/nl/topicus/whighstock/components/WHighStockJavaScriptResourceReference.java
@@ -4,14 +4,14 @@ import org.odlabs.wiquery.core.resources.WiQueryJavaScriptResourceReference;
 
 public class WHighStockJavaScriptResourceReference extends WiQueryJavaScriptResourceReference
 {
-	private static final long serialVersionUID = -4771815414204892357L;
+	private static final long serialVersionUID = -3102442307594519480L;
 
 	private static WHighStockJavaScriptResourceReference INSTANCE =
 		new WHighStockJavaScriptResourceReference();
 
 	private WHighStockJavaScriptResourceReference()
 	{
-		super(WHighStockJavaScriptResourceReference.class, "highstock.src.js");
+		super(WHighStockJavaScriptResourceReference.class, "jquery.highstock.js");
 	}
 
 	public static WHighStockJavaScriptResourceReference get()
