+++ src/main/resources/nl/topicus/whighstock/components/jquery.highstock.js
