+++ src/main/resources/nl/topicus/whighstock/components/themes/jquery.highstock.skies.js
